Track application size state.
